move to global.h, add ctors
