+++ section/include/moho.h
@@ -47,29 +47,6 @@ struct vtable
 	void* methods[];
 };
 
-#define SSO_bytes 0x10ul
-
-template<typename T>
-struct basic_string
-{
-	static constexpr uint32_t sso_size = SSO_bytes/sizeof(T);
-	uint32_t ptr;  // ?
-	T str[sso_size]; // pointer to data
-	uint32_t strLen;
-	uint32_t size; // capacity?
-
-	const T* data() {
-		return size < sso_size ? &str : *(const T**)str;
-	}
-};
-
-using string = basic_string<char>;
-using wstring = basic_string<wchar_t>;
-
-VALIDATE_SIZE(string, 0x1C)
-static_assert(string::sso_size == 0x10);
-VALIDATE_SIZE(wstring, 0x1C)
-static_assert(wstring::sso_size == 0x8);
 
 template<typename T>
 struct vector
